changed to hardlinks
Reversed decision to try to use HTML header for previous URL for redirects. The last URL was nice, but it broke when the form failed or encountered any other issue. This could have been prevented with modals for creation or editing, so that it wouldn't be an issue. Blade needs refactoring, so if I were to redo this. I would use modals.